Capstone1/requirements.txt
streamlit>=1.36.0
google-generativeai==0.8.5
python-dotenv>=1.0.1
requests>=2.31.0